requirements.txt
Django==3.1
gunicorn==20.0.4
djangorestframework
django-filter
